[fix] modified regexp to properly separate app name from its bracketing.
Prevents mismatches when pushing multiple app at the same time where an appname is the prefix of a differet appname in that set.
Before the change the tailer for the shorter appname picked up not only its own pods, but also the pod with the longer name.

IOW the old ".*NAME.*" picks up "NAME" and "NAMELONG".