[feat] Added `Origin` field to log entries, and reconfigured tailer to use.
The origin field is derived from the pod name to summary where the log entry comes from.
At the moment this is staging or app.

This reduces most log entries to something more sensible to user, even with the container name still present.

There are still log entries falling out of this.
These have a green-colored `+ ` as prefix.